Fixed bugs:
rename wasn't working because it was trouncing a container and a container is
a non-empty directory.  Explicitly catch this case and blow away the container
so that the rename can work

untar wasn't working for files with --r--r--r permissions bec first it was
creating the container with --r--r--r and then trying to write to it and the
write was failing w/ a permissions error.  This seems like a problem with tar
but I just removed the access check in f_open because we have users who use
it like a normal file system and they expect tar to work.  They prolly also
expect permissions to work so hopefully they still do....



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs@14 979fb63a-54b4-4514-8d67-9ab8ea1b59df