OK.  It's pretty stable.  Turning off direct_io makes things work well. I can build plfs and I can build the plfs paper.
Two current issues.  One resolved, one open.

One) Resolved.  Right now we create a Plfs_fd in the open and inside
it is one index, and potentially multiple data files per pid.  Writes
specify a pid and that's how they get a data file fd.  For some reason,
in a Makefile in the plfs paper, there is an operation like this:
./get_data > data
And what was happening is that the f_open on the data file was coming
in with one pid, but then the write was coming in with another and 
we were erroring out since that pid wasn't registered in the plfs_fd 
structure.  But since this is something that FUSE apparently does, 
now we handle this by giving an unknown pid some other pid's fd.

Two) Open.  An existing file can be opened with vim and rewritten once.
But then trying to do it a second time fails.  Killing plfs and restarting
then allows it to be rewritten once again.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/library@39 979fb63a-54b4-4514-8d67-9ab8ea1b59df