so previously I'd added some code when I was making the test case work where I was building the plfs SC paper.  That code had this weird thing where the Makefile was opening a file and then writing to it with a child.  I added code so that when that write came in, the child would try to just use the parent's datafile.
But then I found out that was breaking mpi jobs
running over fuse because they were sharing a
datafile but putting their own pids in the index,
so the read was trying to get data from datafiles
that didn't exist

So now it does the thing which it originally did,
which is to have every writer create their own
datafile but share an index file

So now a fs_test with two writers in N-1 mode
works again but I want to test making the plfs
SC paper.  But I can't on this machine since 
it's missing fig2dev.  So I'm gonna commit, 
and pull on a machine with fig2dev.

Also, I tried to do a cvs co plfs from the
old LANL plfs repository and a couple of times
it failed because cvs was trying to rename an
open file and PLFS currently returns -ENOSYS
when that happens.  But most recently that worked
so that might just be a race condition.

Ugh, not sure how to deal with that though
if it reappears.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/library@74 979fb63a-54b4-4514-8d67-9ab8ea1b59df