OK.
On cuda, I was testing mpirun -np 2 fs_test -type 2 -strided 1.
I saw a problem where the 2nd open was using the fd acquired
by the 1st open.  This is broken and then when we tried to 
read the file, we saw writes done by the 2nd open where the index
entry specified the pid of the 2nd open but that datafile didn't
exist so the index creation for the read failed.

The reason that was broken was because when I was doing tests
building the plfs SC paper, there's a part of the build process
that does: ./get_data > data.  And what happens there is that
a parent opens data, and then a child writes to it.  So I added
code where when you ask for an fd in the WriteFile from a pid
and that pid doesn't have an fd, it just gives you another fd that
has been previously registered.

So now I make it where if a 2nd open comes in and doesn't find an
fd, it makes a new one.  This fixed the fs_test.

And if a write comes in with a new pid, it makes a new one as well.
This fixes the ./get_data > data.

Now I need to go back to cuda and make sure that I didn't rebreak
the mpi job.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/library@75 979fb63a-54b4-4514-8d67-9ab8ea1b59df