Executables staged in PLFS cannot be run when PLFS is running with -o direct_io since direct IO skips the buffer cache and therefore the mmap that is needed to exec a file is not possible.
Unfortunately, the error received when this happens is very cryptic.
To fix this, when PLFS detects that it is running in direct_io mode,
it modifies the stat structure to turn off the execute bits on the
get_attr call.  This way, the operating system itself won't try to
exec the file and the user will see an obvious error when they 
chmod to 0700 and they see that the file is 0600.  We actually allow 
the file to be created with the appropriate exec bits but we just 
modify it on the stat.  This way if a user creates an executable file
when direct_io is on, it will be created correctly but just not 
executable.  Then if they unmount and remount w/out direct_io, then 
the file will already have the expected permissions.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/library@77 979fb63a-54b4-4514-8d67-9ab8ea1b59df