got this working.  this branch has new plfsrc directives:
1) canonical_backends
2) shadow_backends

This allows the user to specify that some set of backends are where canonical
containers should be placed and some different set of backends are where
shadows should be placed.  If these directives are used, then shadow subdirs
will ALWAYS be created in shadow backends.  This can be used in instances 
where the canonical is desired to be on slow global and the data pieces are
desired to be on fast local



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/ssdn@663 979fb63a-54b4-4514-8d67-9ab8ea1b59df