insert_mount_point wasn't thread safe due to static set<string> backends which is used to protect against bad plfsrc which has same backend in multiple mount points.  It was static bec it needs to persist across multiple calls. It gets called for each mount point in the plfsrc from parse_conf.  Now parse_conf creates one on stack and passes it.
Was thread problem bec there is a race to create
the plfs_conf at init time for multithreaded 
programs using plfs.  It's fine for there to be
a race.  I suppose we could also deal with this
with a mutex.  That'd probably be safer.  But the
race is fine.  It just happens at init time.  And
as long as all parsing temporary variables are
local (which they weren't but are now) then no
problem.



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/ssdn@683 979fb63a-54b4-4514-8d67-9ab8ea1b59df