a bit more progress towards flat file mode
I got things a bit more set up.  Got stuck on plfs_rename which is tricky
since it may be operating on a logical directory or a logical file.

I still haven't figured out how to get the function pointers in.

for many of the ops, they can be stashed in pconf->mnt_pt but in others
they will need to be stashed in Plfs_Fd



git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/branches/ssdn@710 979fb63a-54b4-4514-8d67-9ab8ea1b59df